requirements.txt
tensorflow==2.13.0
numpy==1.22.4
fastapi==0.97.0
Pillow==9.5.0
python-multipart==0.0.6
uvicorn==0.23.1